tons of improvements towards azd
