+++ .editorconfig
@@ -18,6 +18,11 @@ dotnet_style_qualification_for_method = true
 
 #### Diagnostic configuration ####
 
+# SKEXP0001: Warning: SK Core is experimental
 dotnet_diagnostic.SKEXP0001.severity = none
+# SKEXP0020: Warning: SK Memory connectors are experimental
 dotnet_diagnostic.SKEXP0020.severity = none
+# SKEXP0010: Warning: SK OpenAI is experimental
 dotnet_diagnostic.SKEXP0010.severity = none
+# CA2007: Consider calling ConfigureAwait on the awaited task
+dotnet_diagnostic.CA2007.severity = none
