+++ samples/marketing/infra/modules/fetch-container-image.bicep
@@ -1,0 +1,8 @@
+param exists bool
+param name string
+
+resource existingApp 'Microsoft.App/containerApps@2023-05-02-preview' existing = if (exists) {
+  name: name
+}
+
+output containers array = exists ? existingApp.properties.template.containers : []
