+++ samples/marketing/src/backend/SignalRHub/SignalRConnectionsDB.cs
@@ -3,7 +3,7 @@
 namespace Marketing.SignalRHub;
 public static class SignalRConnectionsDB
 {
-    public static ConcurrentDictionary<string, string> ConnectionIdByUser { get; set; } = new ConcurrentDictionary<string, string>();
-    public static ConcurrentDictionary<string, string> AllConnections { get; set; } = new ConcurrentDictionary<string, string>();
+    public static ConcurrentDictionary<string, string> ConnectionIdByUser { get; } = new ConcurrentDictionary<string, string>();
+    public static ConcurrentDictionary<string, string> AllConnections { get; } = new ConcurrentDictionary<string, string>();
 
 }
