+++ samples/marketing/src/frontend/next.config.mjs
@@ -1,4 +1,16 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  images: {
+    domains: ['dalleproduse.blob.core.windows.net'],
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'dalleproduse.blob.core.windows.net',
+        port: '**',
+        pathname: '**',
+      },
+    ],
+  },
+};
 
 export default nextConfig;
