+++ samples/marketing/src/frontend/src/app/marketing/costs/cost.tsx
@@ -13,9 +13,14 @@ import WorkspacePremiumTwoToneIcon from '@mui/icons-material/WorkspacePremiumTwo
 import GavelIcon from '@mui/icons-material/Gavel';
 import { styled } from '@mui/material/styles';
 import { green, pink } from '@mui/material/colors';
-import AppShortcut from '@mui/icons-material/AttachMoney';
-import LoopIcon from '@mui/icons-material/Loop';
-import { Paper, Card, CardContent, CardHeader, Typography } from '@mui/material';
+import AttachMoneyIcon from '@mui/icons-material/AttachMoney';
+
+const data = [
+  { icon: <HandshakeTwoToneIcon sx={{ color: green[500] }} />, label: 'IPA discount form 9-9 - eur 15k' },
+  { icon: <GavelIcon  sx={{ color: green[500] }}  />, label: '2x1 brewery birthday - eur 250k' },
+  { icon: <HandshakeTwoToneIcon  sx={{ color: green[500] }}  />, label: 'Summer day 1 - CHF 180k' },
+  { icon: <HandshakeTwoToneIcon  sx={{ color: pink[500] }}  />, label: 'Worldcup promo 1.5M' },
+];
 
 const FireNav = styled(List)<{ component?: React.ElementType }>({
   '& .MuiListItemButton-root': {
@@ -31,16 +36,9 @@ const FireNav = styled(List)<{ component?: React.ElementType }>({
   },
 });
 
-type CommunityManagerProps = {
-  article: string;
-  setArticle: React.Dispatch<React.SetStateAction<string>>;
-  open: boolean;
-  setOpen: React.Dispatch<React.SetStateAction<boolean>>;
-  imgUrl: string;
-};
-
-export default function CommunityManager({ article, setArticle, open, setOpen, imgUrl }: CommunityManagerProps) {
-  console.log(`[CommunityManager] Rendering.`);
+export default function CostList() {
+  const [open, setOpen] = React.useState(false);
+  console.log(`[Marketing] Rendering.`);
 
   return (
     <Box sx={{ display: 'flex' }}>
@@ -61,18 +59,18 @@ export default function CommunityManager({ article, setArticle, open, setOpen, i
               '&:hover, &:focus': { '& #arrowdownicon': { opacity: open ? 1 : 0 } },
             }}
           >
-            <ListItemIcon sx={{ my: 0, opacity: 1, class: "menuicon" }}>
-              <AppShortcut />
+            <ListItemIcon sx={{ my: 0, opacity: 1,  class: "menuicon"}}>
+              <AttachMoneyIcon/>
             </ListItemIcon>
             <ListItemText
-              primary="Social Media posts"
+              primary="Economy of similar cases" 
               primaryTypographyProps={{
                 fontSize: 15,
                 fontWeight: 'medium',
                 lineHeight: '20px',
                 mb: '2px',
               }}
-              secondary="Posts in social media"
+              secondary="Cost of similar cases in the past"
               secondaryTypographyProps={{
                 noWrap: true,
                 fontSize: 12,
@@ -91,42 +89,21 @@ export default function CommunityManager({ article, setArticle, open, setOpen, i
               }}
             />
           </ListItemButton>
-          {open && (
-            article === '' || article === null ? (
-              <Box>
-                <LoopIcon
-                  sx={{
-                    animation: "spin 2s linear infinite",
-                    "@keyframes spin": {
-                      "0%": {
-                        transform: "rotate(360deg)",
-                      },
-                      "100%": {
-                        transform: "rotate(0deg)",
-                      },
-                    },
-                  }}
+          {open &&
+            data.map((item) => (
+              <ListItemButton
+                key={item.label}
+                sx={{ py: 0, minHeight: 32, color: 'rgba(255,255,255,.8)' }}
+              >
+                <ListItemIcon sx={{ color: 'inherit' }}>
+                  {item.icon}
+                </ListItemIcon>
+                <ListItemText
+                  primary={item.label}
+                  primaryTypographyProps={{ fontSize: 14, fontWeight: 'medium' }}
                 />
-              </Box>
-            ) : (
-              <Card>
-                <CardContent>
-                    <Typography variant="h5" component="div">
-                      Social media posts on X
-                    </Typography>
-                    <p>{article}</p>
-                    <p style={{ width: '100%', height: '100%' }}>
-                      <img 
-                        src={imgUrl} 
-                        alt="placeholder" 
-                        style={{ width: '100%', height: 'auto' }}/>
-                    </p>
-                    <p>{imgUrl}</p>
-                  </CardContent>
-
-              </Card>
-            )
-          )}
+              </ListItemButton>
+            ))}
         </Box>
       </FireNav>
     </Box>
